AI/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pillow
elasticsearch==8.14.0
sentence-transformers
torch
transformers
xgboost
google-genai
python-multipart
httpx